Feature/unlinked transactions (#133)
* Added linked column

* Changed end dates to end of month for today

* Added new transaction dialog

* Updated tests

* Added check for cannot lock until linked

* Added statement input to new_txn and updated similar matching

* Updated tests

* Fixed pyright issues

+++ tests/models/test_base_uri.py
@@ -70,7 +70,7 @@ class TestBaseURI(TestBase):
         total = n * 8
         self.assertEqual(sum(counts), total)
 
-        std = np.std(counts) / total
+        std = float(np.std(counts) / total)
         self.assertLessEqual(std, 0.05)
 
     def test_table_ids(self) -> None:
